Change to using Tab class and Web Locks API
This is a big refactor and breaking change, so this will be a major version bump. It now requires the Web Locks API which is a bit newer but still widely support. Locks give us very reliable leadership election, allowing the browser to queue requests for leadership and only allow one at a time.